Move KeyMap to sapling::config::KeyMap
